+++ OpenGL-2D/position.h
@@ -8,10 +8,10 @@ class Position
 
 public:
 	Position();
-	Position(float x, float y);
+	Position(int x, int y);
 	~Position();
-	void setX(float x);
-	void setY(float y);
+	void setX(int x);
+	void setY(int y);
 	int getX();
 	int getY();
 	bool operator== (Position& other);
